Move AB test types
