+++ support-frontend/assets/helpers/tracking/trackingOphan.ts
@@ -1,8 +1,9 @@
 // ----- Imports ----- //
 import * as ophan from 'ophan';
-import { type Participations, testIsActive } from 'helpers/abTests/abtest';
+import { testIsActive } from 'helpers/abTests/abtest';
 import { getLocal, setLocal } from 'helpers/storage/storage';
 import type { ReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
+import type { Participations } from '../abTests/models';
 
 // ----- Types ----- //
 // These are to match Thrift definitions which can be found here:
