+++ support-frontend/assets/pages/[countryGroupId]/landingPage.tsx
@@ -1,13 +1,11 @@
-import {
-	getAmountsTestVariant,
-	type Participations,
-} from 'helpers/abTests/abtest';
+import { getAmountsTestVariant } from 'helpers/abTests/abtest';
 import { Country } from 'helpers/internationalisation/classes/country';
 import type { IsoCountry } from 'helpers/internationalisation/country';
 import { type GeoId, getGeoIdConfig } from 'pages/geoIdConfig';
 import { threeTierCheckoutEnabled } from 'pages/supporter-plus-landing/setup/threeTierChecks';
 import { ContributionsOnlyLanding } from 'pages/supporter-plus-landing/twoStepPages/contributionsOnlyLanding';
 import { ThreeTierLanding } from 'pages/supporter-plus-landing/twoStepPages/threeTierLanding';
+import type { Participations } from '../../helpers/abTests/models';
 
 type Props = {
 	geoId: GeoId;
