+++ support-frontend/assets/pages/subscriptions-landing/components/subscriptionsProduct.tsx
@@ -1,9 +1,9 @@
 import cx from 'classnames';
 import type { ReactNode } from 'react';
 import SubscriptionsProductDescription from 'components/subscriptionsProductDescription/subscriptionsProductDescription';
-import type { Participations } from 'helpers/abTests/abtest';
 import 'helpers/types/option';
 import type { ProductButton } from 'pages/subscriptions-landing/copy/subscriptionCopy';
+import type { Participations } from '../../../helpers/abTests/models';
 
 type PropTypes = {
 	title: string;
