+++ support-frontend/assets/pages/subscriptions-landing/copy/subscriptionCopy.tsx
@@ -4,7 +4,6 @@ import DigitalPackshotHero from 'components/packshots/digital-packshot-hero';
 import GuardianWeeklyPackShotHero from 'components/packshots/guardian-weekly-packshot-hero';
 import PaperPackshot from 'components/packshots/paper-packshot';
 // images
-import type { Participations } from 'helpers/abTests/abtest';
 import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
 import { GBPCountries } from 'helpers/internationalisation/countryGroup';
 import { currencies, detect } from 'helpers/internationalisation/currency';
@@ -23,6 +22,7 @@ import {
 	guardianWeeklyLanding,
 	paperSubsUrl,
 } from 'helpers/urls/routes';
+import type { Participations } from '../../../helpers/abTests/models';
 import type { PriceCopy, PricingCopy } from '../subscriptionsLandingProps';
 
 // types
