+++ support-frontend/assets/pages/subscriptions-landing/subscriptionsLandingProps.ts
@@ -1,11 +1,11 @@
 // ----- Imports ----- //
-import type { Participations } from 'helpers/abTests/abtest';
 import { getGlobal } from 'helpers/globalsAndSwitches/globals';
 import { CountryGroup } from 'helpers/internationalisation/classes/countryGroup';
 import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
 import type { SubscriptionProduct } from 'helpers/productPrice/subscriptions';
 import type { ReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
 import { getReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
+import type { Participations } from '../../helpers/abTests/models';
 
 export type PriceCopy = {
 	price: number;
