+++ support-frontend/assets/pages/subscriptions-redemption/api.ts
@@ -1,4 +1,3 @@
-import type { Participations } from 'helpers/abTests/abtest';
 import { fetchJson } from 'helpers/async/fetch';
 import { appropriateErrorMessage } from 'helpers/forms/errorReasons';
 import { postRegularPaymentRequest } from 'helpers/forms/paymentIntegrations/readerRevenueApis';
@@ -27,6 +26,7 @@ import type { Option } from 'helpers/types/option';
 import { routes } from 'helpers/urls/routes';
 import { getOrigin } from 'helpers/urls/url';
 import type { Stage } from 'pages/subscriptions-redemption/subscriptionsRedemptionReducer';
+import type { Participations } from '../../helpers/abTests/models';
 
 type ValidationResult = {
 	valid: boolean;
