+++ support-frontend/assets/pages/supporter-plus-landing/setup/threeTierChecks.ts
@@ -1,6 +1,6 @@
-import type { Participations } from 'helpers/abTests/abtest';
 import type { SelectedAmountsVariant } from 'helpers/contributions';
 import { isContributionsOnlyCountry } from 'helpers/contributions';
+import type { Participations } from '../../../helpers/abTests/models';
 
 export const threeTierCheckoutEnabled = (
 	abParticipations: Participations,
