+++ support-frontend/assets/pages/weekly-subscription-landing/weeklySubscriptionLandingProps.ts
@@ -1,4 +1,3 @@
-import type { Participations } from 'helpers/abTests/abtest';
 import {
 	getGlobal,
 	getProductPrices,
@@ -10,6 +9,7 @@ import type { IsoCountry } from 'helpers/internationalisation/country';
 import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
 import type { ProductPrices } from 'helpers/productPrice/productPrices';
 import type { PromotionCopy } from 'helpers/productPrice/promotions';
+import type { Participations } from '../../helpers/abTests/models';
 
 export type WeeklyLandingPropTypes = {
 	countryId: IsoCountry;
